+++ src/main/java/uk/gov/pay/publicauth/exception/TokenRevokedExceptionMapper.java
@@ -1,6 +1,6 @@
 package uk.gov.pay.publicauth.exception;
 
-import uk.gov.pay.commons.model.ErrorIdentifier;
+import uk.gov.service.payments.commons.model.ErrorIdentifier;
 
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
