+++ src/main/java/uk/gov/pay/publicauth/filters/LoggingMDCRequestFilter.java
@@ -7,7 +7,7 @@ import javax.ws.rs.container.ContainerRequestFilter;
 import java.io.IOException;
 import java.util.Optional;
 
-import static uk.gov.pay.logging.LoggingKeys.GATEWAY_ACCOUNT_ID;
+import static uk.gov.service.payments.logging.LoggingKeys.GATEWAY_ACCOUNT_ID;
 
 public class LoggingMDCRequestFilter implements ContainerRequestFilter {
