+++ src/main/java/uk/gov/pay/publicauth/filters/LoggingMDCResponseFilter.java
@@ -7,7 +7,7 @@ import javax.ws.rs.container.ContainerResponseContext;
 import javax.ws.rs.container.ContainerResponseFilter;
 import java.io.IOException;
 
-import static uk.gov.pay.logging.LoggingKeys.GATEWAY_ACCOUNT_ID;
+import static uk.gov.service.payments.logging.LoggingKeys.GATEWAY_ACCOUNT_ID;
 
 public class LoggingMDCResponseFilter implements ContainerResponseFilter {
     @Override
