+++ src/test/java/uk/gov/pay/publicauth/it/PublicAuthResourceIT.java
@@ -36,8 +36,8 @@ import static org.hamcrest.Matchers.lessThan;
 import static org.hamcrest.Matchers.not;
 import static org.hamcrest.Matchers.notNullValue;
 import static org.hamcrest.core.Is.is;
-import static uk.gov.pay.commons.model.ErrorIdentifier.AUTH_TOKEN_INVALID;
-import static uk.gov.pay.commons.model.ErrorIdentifier.AUTH_TOKEN_REVOKED;
+import static uk.gov.service.payments.commons.model.ErrorIdentifier.AUTH_TOKEN_INVALID;
+import static uk.gov.service.payments.commons.model.ErrorIdentifier.AUTH_TOKEN_REVOKED;
 import static uk.gov.pay.publicauth.model.TokenPaymentType.CARD;
 import static uk.gov.pay.publicauth.model.TokenPaymentType.DIRECT_DEBIT;
 import static uk.gov.pay.publicauth.model.TokenSource.API;
